requirements.txt
Pillow>=9.0.0
PySide6>=6.2.0
pyinstaller>=5.0
